DO NOT MERGE: manifest: add support for extended vold/fs compatibility
I've imported CyanogenMod changes to support misc filesytems and minivold as
we had in marshmallow-caf.
Those versions contain major updates on top of CAF to allow booting on a
wider range of devices, F2FS and BoringSSL crypto.

Repos to be created:
https://github.com/GalaticStryder/external_e2fsprogs
https://github.com/GalaticStryder/external_gptfdisk

Branches to be pushed:
https://github.com/GalaticStryder/external_f2fs-tools
https://github.com/GalaticStryder/system_vold
https://github.com/GalaticStryder/system_extras

PS1: Remove system/media, not needed. The commit will be cleaned once we
push them and it's ready 2go.

Change-Id: Ic09fc9e54bfa81bf5feb99a52e2130a4560006ff
Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>