pa: Paranoid Android rebrand (DO NOT MERGE, JUST TEMPLATE)
<!-- This is just to save as a template and it´s waiting on Arz future approval.-->

The main idea is to go fully Paranoid Android having AOSPA as a backup,
but don´t have the same thing under two names, which makes it confusing
to some users and even media people.

As Paranoid Android has been there since its start, and now that we are
full CAF based, I think it´s time to give one more step and let AOSPA
rests while it´s on his climax.

Actually most of the ROM stuff is Paranoid Android everywhere, even the
wallpapers and the only AOSPA stuff is actually the ending part of the
bootanimation. Server side, we just need to flip on using paranoidandroid.co
on official post releases and leave aospa.co acting as a backup link as
the other one is acting right now. This plus the new website would come
too good for the Android O release.

This would help on getting a more complete branding, easier to find,
easier to recognize and be Paranoid!

PD: In case this ever gets forward, here we have the first commit for
do the rebranding. All the decision is on Arz hands, just leaving it
there in case someone wants to propose or show his ideas.

Change-Id: I9bf3698f4705ceddf26c88f84a1a87c8adf73749