Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I'm one of the people who still thinks GCC is better for kernels, and I use GCC for all of my kernels.
> 
> But I'm not sure whether we should track it globally or make this a device-specific roomservice.

I'm fine with doing it globally, in general, it would be better to have a standard compiler if we can (doesn't have to be a requirement). What's the level of effort to support this gcc version on other devices?

Patch-set: 1
Reviewer: Gerrit User 1000012 <1000012@3162e3d8-d894-470e-abd9-724f7b0effad>