Update patch set 2
Patch Set 2:

Honestly the NFC stack is so bad, is it even worth building from source? Random tags break everything, some branches work on different hardware, should we just use blobs?

If not, I'm fine with this change, just need to let Maintainers using the OSS stack know. OP6 and OP7 use the NXP non-Qualcomm stack, which further adds to the mess that is NFC.

Patch-set: 2
CC: Gerrit User 1000012 <1000012@3162e3d8-d894-470e-abd9-724f7b0effad>